Refactor Taxas tab into dedicated screen
